+++ bench/type4/adversarial/cases/swift_flat_map_custom_overload/consumer.swift
@@ -1,0 +1,3 @@
+func flatMapCustomOverload(_ groups: [[Bool]]) -> [Bool] {
+    groups.flatMap { (group: [Bool]) in group.map { value in value } }
+}